Fix for issue #25
The "Disabled" button is not really disabled while editing an instance. Users might click on it and it will actually toggle the open property. Submitting an event after clicking once the Disabled button would actually submit rec_type info and create a series that has a parent, which is wrong.