Make it possible to target expanded scheduler with CSS selectors
Currently it is not possible to apply a different style to the scheduler based on the state of the expand plugin via CSS selectors, because all changes are made by modifying style properties directly. This change will add `dhx_scheduler_expanded` class every time the scheduler is expanded and remove it, when the scheduler is collapsed. I didn't use the `classList` property, so it should work even with older browsers.